requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.12=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
argh=0.26.2=py38_0
asgiref=3.5.1=pypi_0
astroid=2.6.6=py38haa95532_0
async_generator=1.10=pyhd3eb1b0_0
atomicwrites=1.4.0=py_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.6=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
bcrypt=3.2.0=py38he774522_0
black=19.10b0=py_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py38h2a96729_1
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py38haa95532_0
cffi=1.14.6=py38h2bbff1b_0
chardet=4.0.0=py38haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.1=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.8=py38h71e12ea_0
cycler=0.10.0=py38_0
cython=0.29.24=py38hd77b12b_0
debugpy=1.4.1=py38hd77b12b_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=pyhd3eb1b0_0
django=4.0.4=pypi_0
docutils=0.17.1=py38haa95532_1
ecdsa=0.18.0=pypi_0
entrypoints=0.3=py38_0
flake8=3.9.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py38_1
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.2=pyhd3eb1b0_0
image=1.5.33=pypi_0
imagesize=1.2.0=pyhd3eb1b0_0
importlib-metadata=4.8.1=py38haa95532_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=haa95532_3372
intervaltree=3.1.0=pyhd3eb1b0_0
ipykernel=6.2.0=py38haa95532_1
ipython=7.27.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.9.3=pyhd3eb1b0_0
jedi=0.17.2=py38haa95532_1
jinja2=3.0.1=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_core=4.7.1=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
keyring=23.1.0=py38haa95532_0
kiwisolver=1.3.1=py38hd77b12b_0
lazy-object-proxy=1.6.0=py38h2bbff1b_0
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libspatialindex=1.9.3=h6c2663c_0
libtiff=4.2.0=hd0e1b90_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.0.1=py38h2bbff1b_0
matplotlib=3.4.2=py38haa95532_0
matplotlib-base=3.4.2=py38h49ac443_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=py38_1
mistune=0.8.4=py38he774522_1000
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.0=py38h277e83a_2
mkl_random=1.2.2=py38hf11a4ad_0
mpmath=1.2.1=py38haa95532_0
munkres=1.1.4=py_0
mypy_extensions=0.4.3=py38_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
numexpr=2.7.3=py38hb80d3ca_1
numpy=1.20.3=py38ha4e8547_0
numpy-base=1.20.3=py38hc2deb75_0
numpydoc=1.1.0=pyhd3eb1b0_1
olefile=0.46=pyhd3eb1b0_0
opencv=4.0.1=py38h2a7c758_0
opencv-contrib-python=4.5.3.56=pypi_0
openssl=1.1.1s=h2bbff1b_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.3=py38h6214cd6_0
pandocfilters=1.4.3=py38haa95532_1
paramiko=2.7.2=py_0
parso=0.7.0=py_0
pathspec=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py38h4fa10fc_0
pip=21.2.2=py38haa95532_0
pluggy=0.13.1=py38haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.8.0=py38h2bbff1b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=4.0.1=py38he44ac1e_0
pycodestyle=2.6.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pydocstyle=6.1.1=pyhd3eb1b0_0
pyflakes=2.2.0=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pylint=2.9.6=py38haa95532_1
pyls-black=0.4.6=hd3eb1b0_0
pyls-spyder=0.3.2=pyhd3eb1b0_0
pynacl=1.5.0=py38h8cc25b3_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.17.3=py38he774522_0
pysocks=1.7.1=py38haa95532_0
python=3.8.11=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.36.2=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pywin32=228=py38hbaba5e8_1
pywin32-ctypes=0.2.0=py38_1000
pyyaml=5.4.1=py38h2bbff1b_1
pyzmq=22.2.1=py38hd77b12b_1
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.2=pyhd3eb1b0_0
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
regex=2021.8.3=py38h2bbff1b_0
requests=2.26.0=pyhd3eb1b0_0
rope=0.19.0=pyhd3eb1b0_0
rtree=0.9.7=py38h2eaa2aa_1
scipy=1.7.1=py38hbe87c03_2
setuptools=58.0.4=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.16.0=pyhd3eb1b0_0
snowballstemmer=2.1.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sphinx=4.2.0=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
spyder=4.2.5=py38haa95532_0
spyder-kernels=1.10.2=py38haa95532_0
sqlite=3.36.0=h2bbff1b_0
sqlparse=0.4.2=pypi_0
sympy=1.8=py38haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
textdistance=4.2.1=pyhd3eb1b0_0
three-merge=0.1.1=pyhd3eb1b0_0
tk=8.6.11=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.1.0=pyhd3eb1b0_0
typed-ast=1.4.3=py38h2bbff1b_1
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2022.1=pypi_0
ujson=4.0.2=py38hd77b12b_0
urllib3=1.26.6=pyhd3eb1b0_1
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=1.0.2=py38haa95532_1
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
wheel=0.37.0=pyhd3eb1b0_1
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
wrapt=1.12.1=py38he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
